refactor(core): split oversized recall modules
Extract benchmark support scoring, composite profile storage, and query-aware recall prefiltering into focused modules. Preserve behavior while bringing the grandfathered file-size ratchets back under their enforced ceilings.
