+++ packages/remnic-core/src/orchestration/recall-search-pipeline.ts
@@ -34,7 +34,13 @@ import { RerankCache, rerankLocalOrNoop, reorderByRankedKeys } from "../rerank.j
 import type { SearchBackend, SearchDegradation, SearchExecutionOptions, SearchQueryOptions } from "../search/port.js";
 import { SecureStoreLockedError } from "../secure-store/index.js";
 import { isPathInsideStorageRoot } from "../storage-paths.js";
-import { extractTagsFromPrompt, isTemporalQuery, queryByDateRangeAsync, queryByTagsAsync, readIndexSnapshotAsync, recencyWindowFromPrompt, resolvePromptTagPrefilterAsync } from "../temporal-index.js";
+import { extractTagsFromPrompt, isTemporalQuery, queryByDateRangeAsync, queryByTagsAsync, readIndexSnapshotAsync, recencyWindowFromPrompt } from "../temporal-index.js";
+import {
+  buildQueryAwarePrefilter as buildQueryAwarePrefilterHelper,
+  fetchActiveArtifactsForNamespace as fetchActiveArtifactsForNamespaceHelper,
+  searchEmbeddingFallback as searchEmbeddingFallbackHelper,
+  searchQueryAwareFallback as searchQueryAwareFallbackHelper,
+} from "./recall-search-prefilter.js";
 import { shouldFilterSupersededFromRecall } from "../temporal-supersession.js";
 import { isValidAsOf, isValidityExpiredNow } from "../temporal-validity.js";
 import type { TrustStageResultItem } from "../trust-score-stage.js";
@@ -49,7 +55,6 @@ import {
 } from "../recall-memory-map.js";
 import { type UtilityRuntimeValues, applyUtilityRankingRuntimeDelta } from "../utility-runtime.js";
 import {
-  computeArtifactCandidateFetchLimit,
   lifecycleRecallScoreAdjustment,
   shouldFilterLifecycleRecallCandidate,
   computeQmdHybridFetchLimit,
@@ -247,157 +252,16 @@ export class RecallSearchPipelineCoordinator {
     prompt: string,
     targetCount: number,
   ): Promise<MemoryFile[]> {
-    const storage = await this.deps.storageRouter.storageFor(namespace);
-    let fetchLimit = computeArtifactCandidateFetchLimit(targetCount);
-    const maxFetchLimit = Math.min(800, Math.max(fetchLimit, targetCount * 8));
-    const MAX_ATTEMPTS = 4;
-    let bestFiltered: MemoryFile[] = [];
-
-    for (let attempt = 0; attempt < MAX_ATTEMPTS; attempt += 1) {
-      const rawResults = await storage.searchArtifacts(prompt, fetchLimit);
-      const sourceIds = Array.from(
-        new Set(
-          rawResults
-            .map((a) => a.frontmatter.sourceMemoryId)
-            .filter(
-              (id): id is string => typeof id === "string" && id.length > 0,
-            ),
-        ),
-      );
-      const sourceStatus =
-        sourceIds.length > 0
-          ? await this.deps.resolveArtifactSourceStatuses(storage, sourceIds)
-          : new Map<string, "active" | "superseded" | "archived" | "missing">();
-
-      const filtered: MemoryFile[] = [];
-      for (const artifact of rawResults) {
-        const sourceId = artifact.frontmatter.sourceMemoryId;
-        if (!sourceId) {
-          filtered.push(artifact);
-          if (filtered.length >= targetCount) break;
-          continue;
-        }
-        const status = sourceStatus.get(sourceId) ?? "missing";
-        if (status !== "active") continue;
-        filtered.push(artifact);
-        if (filtered.length >= targetCount) break;
-      }
-
-      if (filtered.length >= targetCount) return filtered.slice(0, targetCount);
-      if (filtered.length > bestFiltered.length) {
-        bestFiltered = filtered;
-      }
-      if (rawResults.length === 0) return filtered;
-      if (rawResults.length < fetchLimit && filtered.length > 0)
-        return filtered;
-      if (fetchLimit >= maxFetchLimit) return filtered;
-
-      const growth = Math.max(targetCount * 2, 12);
-      fetchLimit = Math.min(maxFetchLimit, fetchLimit + growth);
-    }
-
-    return bestFiltered;
+    return fetchActiveArtifactsForNamespaceHelper(this.deps, namespace, prompt, targetCount);
   }
 
   async buildQueryAwarePrefilter(
     prompt: string,
     recallNamespaces: string[],
   ): Promise<QueryAwarePrefilter> {
-    if (!resolveIndexingCapabilities(this.deps.config).queryAwareIndexing || !prompt.trim()) {
-      return {
-        candidatePaths: null,
-        temporalFromDate: null,
-        matchedTags: [],
-        expandedTags: [],
-        combination: "none",
-        filteredToFullSearch: false,
-      };
-    }
-
-    const temporalFromDate = isTemporalQuery(prompt)
-      ? recencyWindowFromPrompt(prompt, Date.now())
-      : null;
-    // Read the temporal + tag prefilter indexes as ONE consistent snapshot
-    // (issue #1911, Codex Medium) so a concurrent async index mutation can't be
-    // observed half-applied (new temporal row, tag membership not yet written),
-    // which would drop the current memory from the tag prefilter.
-    const [rawTemporal, tagSignals] = await readIndexSnapshotAsync(
-      this.deps.config.memoryDir,
-      () =>
-        Promise.all([
-          temporalFromDate
-            ? queryByDateRangeAsync(this.deps.config.memoryDir, temporalFromDate)
-            : Promise.resolve<Set<string> | null>(null),
-          resolvePromptTagPrefilterAsync(this.deps.config.memoryDir, prompt).catch(
-            () => ({
-              matchedTags: extractTagsFromPrompt(prompt),
-              expandedTags: extractTagsFromPrompt(prompt),
-              paths: null,
-            }),
-          ),
-        ]),
-    );
-
-    const temporalCandidates = this.deps.scopeQueryAwarePaths(
-      rawTemporal,
-      recallNamespaces,
-    );
-    const tagCandidates = this.deps.scopeQueryAwarePaths(
-      tagSignals.paths,
-      recallNamespaces,
-    );
-    const maxCandidates = this.deps.config.queryAwareIndexingMaxCandidates;
-
-    let candidatePaths: Set<string> | null = null;
-    let combination: QueryAwarePrefilter["combination"] = "none";
-    let filteredToFullSearch = false;
-
-    if (
-      tagSignals.matchedTags.length > 0 &&
-      tagCandidates !== null &&
-      tagCandidates.size === 0
-    ) {
-      candidatePaths = tagCandidates;
-      combination = "tag";
-    } else if (temporalCandidates !== null && tagCandidates !== null) {
-      const intersection = new Set(
-        Array.from(temporalCandidates).filter((memoryPath) =>
-          tagCandidates.has(memoryPath),
-        ),
-      );
-      if (intersection.size > 0) {
-        candidatePaths = intersection;
-        combination = "intersection";
-      } else {
-        candidatePaths = new Set([...temporalCandidates, ...tagCandidates]);
-        combination = "union";
-      }
-    } else if (temporalCandidates !== null) {
-      candidatePaths = temporalCandidates;
-      combination = "temporal";
-    } else if (tagCandidates !== null) {
-      candidatePaths = tagCandidates;
-      combination = "tag";
-    }
-
-    if (
-      candidatePaths &&
-      maxCandidates > 0 &&
-      candidatePaths.size > maxCandidates
-    ) {
-      filteredToFullSearch = true;
-      candidatePaths = null;
-    }
-
-    return {
-      candidatePaths,
-      temporalFromDate,
-      matchedTags: tagSignals.matchedTags,
-      expandedTags: tagSignals.expandedTags,
-      combination,
-      filteredToFullSearch,
-    };
+    return buildQueryAwarePrefilterHelper(this.deps, prompt, recallNamespaces);
   }
+
 
   async fetchQmdMemoryResultsWithArtifactTopUp(
     prompt: string,
@@ -654,26 +518,7 @@ export class RecallSearchPipelineCoordinator {
     query: string,
     limit: number,
   ): Promise<QmdSearchResult[]> {
-    if (!resolveMemoryLifecycleCapabilities(this.deps.config).embeddingFallback) return [];
-    if (!(await this.deps.embeddingFallback.isAvailable())) return [];
-    const hits = await this.deps.embeddingFallback.search(query, limit);
-    if (hits.length === 0) return [];
-
-    const results: QmdSearchResult[] = [];
-    for (const hit of hits) {
-      const fullPath = path.isAbsolute(hit.path)
-        ? hit.path
-        : path.join(this.deps.config.memoryDir, hit.path);
-      const memory = await this.deps.storage.readMemoryByPath(fullPath);
-      if (!memory) continue;
-      results.push({
-        docid: hit.id,
-        path: fullPath,
-        score: hit.score,
-        snippet: memory.content.slice(0, 400).replace(/\n/g, " "),
-      });
-    }
-    return results;
+    return searchEmbeddingFallbackHelper(this.deps, query, limit);
   }
 
   /**
@@ -686,23 +531,9 @@ export class RecallSearchPipelineCoordinator {
     queryAwarePrefilter?: QueryAwarePrefilter,
     abortSignal?: AbortSignal,
   ): Promise<QmdSearchResult[]> {
-    throwIfRecallAborted(abortSignal);
-    const cappedLimit = Math.max(0, limit);
-    if (cappedLimit === 0 || queryAwarePrefilter?.candidatePaths?.size === 0) return [];
-
-    const candidatePaths = queryAwarePrefilter?.candidatePaths;
-    const scopedSeedResults = (
-      candidatePaths?.size
-        ? await this.deps.searchScopedMemoryCandidates(
-            candidatePaths,
-            prompt,
-            candidatePaths.size,
-            { allowArchived: true },
-          )
-        : []
-    ).filter((result) => !isGenericRecallExcludedPath(result.path, this.deps.config, "qmd"));
-    return scopedSeedResults.slice(0, cappedLimit);
+    return searchQueryAwareFallbackHelper(this.deps, prompt, limit, queryAwarePrefilter, abortSignal);
   }
+
 
   async applyColdFallbackPipeline(options: {
     prompt: string;
